Busca por similaridade para Patterns e Anti-Patterns  FUNCIONANDO!!!!Quase chorei quando vi!!

+++ src/br/edu/fei/sigepapp/bancodedados/model/AtributosBuscaSimilaridade.java
@@ -4,9 +4,6 @@
  */
 package br.edu.fei.sigepapp.bancodedados.model;
 
-import java.io.Serializable;
-import java.math.BigDecimal;
-
 /**
  *
  * @author lopespt
@@ -14,7 +11,7 @@ import java.math.BigDecimal;
 public class AtributosBuscaSimilaridade {
 
     long cd_objeto;
-    BigDecimal similaridade;
+    double similaridade;
     String tp_estrutura;
     String nm_objeto;
     String contexto;
@@ -25,7 +22,7 @@ public class AtributosBuscaSimilaridade {
     public AtributosBuscaSimilaridade() {
     }
 
-    public AtributosBuscaSimilaridade(long cd_objeto, BigDecimal similaridade,String tp_estrutura, String nm_objeto, String contexto, String problema, String solucao) {
+    public AtributosBuscaSimilaridade(long cd_objeto, double similaridade,String tp_estrutura, String nm_objeto, String contexto, String problema, String solucao) {
         this.cd_objeto = cd_objeto;
         this.similaridade = similaridade;
         this.tp_estrutura=tp_estrutura;
@@ -67,11 +64,11 @@ public class AtributosBuscaSimilaridade {
         this.problema = problema;
     }
 
-    public BigDecimal getSimilaridade() {
+    public double getSimilaridade() {
         return similaridade;
     }
 
-    public void setSimilaridade(BigDecimal similaridade) {
+    public void setSimilaridade(double similaridade) {
         this.similaridade = similaridade;
     }
